+++ scripts/make-react.js
@@ -51,18 +51,11 @@ components.map(component => {
       ${eventImports}
       ${eventExports}
 
-      export type ForwardComponent<
-        Element extends HTMLElement,
-        ReactComponent extends React.ElementType
-      > = React.JSXElementConstructor<
-        React.ComponentPropsWithoutRef<ReactComponent> & {
-          ref?: React.ForwardedRef<Element>;
-        }
-      > & { displayName?: string }
-
       const tagName = '${component.tagName}'
+      Component.define('${component.tagName}')
 
-      const component = createComponent({
+      ${jsDoc}
+      const reactWrapper = createComponent({
         tagName,
         elementClass: Component,
         react: React,
@@ -72,20 +65,7 @@ components.map(component => {
         displayName: "${component.name}"
       })
 
-      ${jsDoc}
-      class SlComponent extends React.Component<Parameters<typeof component>[0]> {
-        constructor (...args: Parameters<typeof component>) {
-          super(...args)
-          Component.define(tagName)
-        }
-
-        render () {
-          const { children, ...props } = this.props
-          return React.createElement(component, props, children)
-        }
-      }
-
-      export default SlComponent as ForwardComponent<Component, typeof SlComponent>;
+      export default reactWrapper
     `,
     Object.assign(prettierConfig, {
       parser: 'babel-ts'